handle special path "/" when making paths relative
In the previous implementation, the leading "/" was naively dropped
generating the empty string—which is of course no valid path—when the
root folder ("/") is passed. This patch handles this special by
replacing "/" by ".".

I addition tests will fail using directory-1.3.3.1, because the empty
file path is not considered to exist anymore. See the changelog

https://hackage.haskell.org/package/directory-1.3.3.1/changelog